feat: switch season with input instead of dropdown
